single-server-agents/pyproject.toml
[project]
name = "single-server-agents"
version = "0.1.0"
description = "This is the rig for stdio mcps with langgraph orchestration"
authors = [
    {name = "The Moose"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "langchain",
    "langgraph",
    "langchain-mcp-adapters",
    "python-dotenv",
    "requests",
    "langchain-openai (>=0.3.11,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "*"
